Second version of combat
